Selling & buying shizzle works

+++ src/F1Manager.Teams/Entities/EngineEntity.cs
@@ -7,9 +7,9 @@ namespace F1Manager.Teams.Entities
     {
         public string Name { get; set; }
         public string Manufacturer { get; set; }
-        public decimal Value { get; set; }
+        public double Value { get; set; }
         public string PictureUrl { get; set; }
-        public decimal WeeklyWearDown { get; set; }
-        public decimal MaximumWearDown { get; set; }
+        public double WeeklyWearDown { get; set; }
+        public double MaximumWearDown { get; set; }
     }
 }
